+++ makeflow/src/dag_node.c
@@ -404,79 +404,6 @@ const struct rmsummary *dag_node_dynamic_label(const struct dag_node *n) {
 	return category_dynamic_task_max_resources(n->category, n->resources_requested, n->resource_request);
 }
 
-struct batch_task *dag_node_to_batch_task(struct dag_node *n, struct batch_queue *queue, int full_env_list)
-{
-	struct batch_task *task = batch_task_create(queue);
-	task->taskid = n->nodeid;
-
-	if(n->type==DAG_NODE_TYPE_WORKFLOW) {
-
-		char *cmd = string_format("makeflow -T local %s",n->workflow_file);
-		char *oldcmd = 0;
-
-		if(n->workflow_is_jx) {
-			oldcmd = cmd;
-			cmd = string_format("%s --jx",cmd);
-			free(oldcmd);
-		}
-
-		if(n->workflow_args) {
-			char args[] = "makeflow.jx.args.XXXXXX";
-			int fd = mkstemp(args);
-			FILE *argsfile = fdopen(fd,"w");
-			jx_print_stream(n->workflow_args,argsfile);
-			fclose(argsfile);
-
-			oldcmd = cmd;
-			cmd = string_format("%s --jx-args %s",cmd,args);
-			free(oldcmd);
-
-			batch_task_add_input_file(task,args,args);
-		}
-
-		if(n->resources_requested->cores>0) {
-			oldcmd = cmd;
-			cmd = string_format("%s --local-cores %d",cmd,(int)n->resources_requested->cores);
-			free(oldcmd);
-		}
-
-		if(n->resources_requested->memory>0) {
-			oldcmd = cmd;
-			cmd = string_format("%s --local-memory %d",cmd,(int)n->resources_requested->cores);
-			free(oldcmd);
-		}
-
-		if(n->resources_requested->disk>0) {
-			oldcmd = cmd;
-			cmd = string_format("%s --local-disk %d",cmd,(int)n->resources_requested->cores);
-			free(oldcmd);
-		}
-
-		batch_task_set_command(task, cmd);
-		batch_task_add_input_file(task,n->workflow_file,n->workflow_file);
-		free(cmd);
-	} else {
-		batch_task_set_command(task, n->command);
-	}
-
-	struct dag_file *f;
-	list_first_item(n->source_files);
-	while((f = list_next_item(n->source_files))){
-		batch_task_add_input_file(task, f->filename, dag_node_get_remote_name(n, f->filename));
-	}
-
-	list_first_item(n->target_files);
-	while((f = list_next_item(n->target_files))){
-		batch_task_add_output_file(task, f->filename, dag_node_get_remote_name(n, f->filename));
-	}
-
-	batch_task_set_resources(task, dag_node_dynamic_label(n));
-
-	batch_task_set_envlist(task, dag_node_env_create(n->d, n, full_env_list));
-
-	return task;
-}
-
 /* Return JX object containing cmd, inputs, outputs, env, and resources. */
 
 struct jx * dag_node_to_jx( struct dag *d, struct dag_node *n, int send_all_local_env)
